# src/main/resources/apispec.yaml
openapi: 3.0.3
info:
  title: OpenApiWeek-5 - OpenAPI 3.0
  description: "zatichka"
  version: "1.0"

servers:
  - url: https://http://localhost:8080/
tags:
  - name: openApiCrud
    description: CRUD API работы с заказами

  - name: order
    description: Access to orders

paths:
  /orders:
    get:
      tags:
        - order
      summary: Find all orders
      description: Returns all order
      operationId: getAllOrders
      parameters:
        - name: status
          in: query
          example: APPROVED
          description: status of order
          required: true
          schema:
            type: string
        - name: startDateTime
          in: query
          description: Date of start
          required: true
          example: 2021-08-11T10:48:01.679Z
          schema:
            type: string
            format: date-time
        - name: endDateTime
          in: query
          description: Date of end
          required: true
          example: 2022-08-10T10:48:01.679Z
          schema:
            type: string
            format: date-time

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orderDTO'
            text/csv:
              { }

        '400':
          description: Invalid ID supplied
          content:
            application/problem+json:
              schema:
                type: string
        '404':
          description: order not found
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string


    post:
      tags:
        - order
      summary: Add a new order
      description: Add a new order
      operationId: addOrder
      requestBody:
        description: Create a new order in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/orderDTO'
        '405':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string
    put:
      tags:
        - order
      summary: Upload order list from csv file
      description: Upload order list from csv file
      operationId: uploadOrderList
      requestBody:
        description: Put order list
        content:
          text/csv:
            schema:
              type: string
              format: binary
        required: true
      responses:
        '202':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/orderDTO'
        '405':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string




  /orders/{orderId}:
    put:
      tags:
        - order
      summary: Update an existing order
      description: Update an existing order by Id
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          description: Id of order to replace
          required: true
          example: 1
          schema:
            $ref: '#/components/schemas/orderId'
      requestBody:
        description: Replace an existent order
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/orderDTO'

        '400':
          description: Invalid ID supplied
        '404':
          description: Order not found
          content:
            application/problem+json:
              schema:
                type: string
        '405':
          description: Validation exception
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string

    get:
      tags:
        - order
      summary: Find order by ID
      description: Returns a single order
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: ID of order to return
          required: true
          example: 1
          schema:
            $ref: '#/components/schemas/orderId'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/orderDTO'
            text/csv:
              { }

        '400':
          description: Invalid ID supplied
          content:
            application/problem+json:
              schema:
                type: string
        '404':
          description: order not found
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string
    patch:
      tags:
        - order
      summary: Update a existing order
      description: Update an existing order by ID
      operationId: patchOrder
      parameters:
        - name: orderId
          in: path
          example: 1
          description: ID of order to return
          required: true
          schema:
            $ref: '#/components/schemas/orderId'
      requestBody:
        description: Update an existent order
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderUpdateDTO'
        required: true
      responses:
        '200':
          description: Succesfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/orderDTO'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                type: string
        '404':
          description: Order not found
          content:
            application/problem+json:
              schema:
                type: string
        '405':
          description: Validation exception
          content:
            application/problem+json:
              schema:
                type: string
        '409':
          description: Illegal State
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string



    delete:
      tags:
        - order
      summary: Deletes a order
      description: delete a order
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          example: 1
          description: order id to delete
          required: true
          schema:
            $ref: '#/components/schemas/orderId'
      responses:
        '200':
          description: Sucess
        '400':
          description: Invalid order value
          content:
            application/problem+json:
              schema:
                type: string
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                type: string







components:
  schemas:
    orderId:
      description: order id to delete
      example: 23
      type: integer
      format: int64
    status:
      type: string
      description: Order Status
      example: APPROVED
      enum:
        - placed
        - approved
        - delivered
    orderDTO:
      description: Order object
      required:
        - productID
        - quantity
      type: object
      properties:
        id:
          type: string
          example: or23
        productID:
          minimum: 1
          type: integer
          format: int64
          example: 198772
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          format: int32
          example: 7
        date:
          type: string
          format: date-time
          example: 2022-08-10T10:48:01.679Z
        status:
          $ref: '#/components/schemas/status'
        complete:
          type: boolean
          example: true
    orderUpdateDTO:
      description: Order update response object
      type: object
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          format: int32
          example: 8
        complete:
          type: boolean
          example: false
        status:
          $ref: '#/components/schemas/status'